feat(invoices): config merge
